[MODIFY] DashboardResponse 값들을 long type -> int type

+++ InTechs-backend/src/main/java/InTechs/InTechs/project/value/IssuesCountInfo.java
@@ -8,8 +8,8 @@ import lombok.ToString;
 @ToString
 @Builder
 public class IssuesCountInfo {
-    long forMe;
-    long resolved;
-    long unresolved;
-    long forMeAndUnresolved;
+    int forMe;
+    int resolved;
+    int unresolved;
+    int forMeAndUnresolved;
 }
